Beginning to look like the problem.
It seems that it's something to do with the array that's passed to
Ember.Select. If I pass an array constructed by finding both authors
separately, it kills the model from the post and won't set new ones,
complaining that "You can only add a 'author' record to this relationship".

I'm not yet sure whether this is because of the array, or because of the
possibility that when the view is rendered the promises haven't resolved yet.